refactor: 스트라이크 존 종류에 따라 재사용 가능하도록 components로 이동
