+++ src/components/InningPlot/InningPlot.tsx
@@ -1,7 +1,7 @@
+import { InningPlotContainer } from '@components/InningPlot/InningPlot.styles';
 import StrikeZone from '@components/StrikeZone/StrikeZone';
+import { StrikeZoneList } from '@components/StrikeZone/StrikeZone.styles';
 
-import { StrikeZoneList } from '@pages/GameDetailPage/GameDetailPage.styles';
-import { InningPlotContainer } from '@pages/GameDetailPage/components/InningPlot/InningPlot.styles';
 import { SectionTitle } from '@pages/GameDetailPage/components/StrikeZoneSection/StrikeZoneSection.styles';
 
 import { AtBat } from '@typings/atbat';
