+++ src/components/SidePlot/SidePlot.styles.tsx
