+++ src/components/SummaryPlot/SummaryPlot.styles.tsx
